Add files via upload
A web app that uses computer vision and generative AI to reward smiles. It is basically created for the students who are currently preparing for placements and may get into self doubt and depression during placement season. This app encourages them to smile and the ai integrated with this interacts with them in order to get their emotions out and making them feel light.